+++ bukkit/src/main/java/us/myles/ViaVersion/bukkit/listeners/protocol1_9to1_8/HandItemCache.java
@@ -34,7 +34,7 @@ public class HandItemCache extends BukkitRunnable {
     }
 
     public static Item convert(ItemStack itemInHand) {
-        if (itemInHand == null) return new Item((short) 0, (byte) 0, (short) 0, null);
-        return new Item((short) itemInHand.getTypeId(), (byte) itemInHand.getAmount(), itemInHand.getDurability(), null);
+        if (itemInHand == null) return new Item(0, (byte) 0, (short) 0, null);
+        return new Item(itemInHand.getTypeId(), (byte) itemInHand.getAmount(), itemInHand.getDurability(), null);
     }
 }
